Order pie charts with biggest category first
The UI shows the label of the first category of a pie chart
in the center of the chart. At a glance this always looks like
it's linked to the biggest colour on the chart. By ordering
so that the biggest category is the first, we ensure that
this visual link makes sense.