Fix usage of d3.map.
As of D3 version 3.4.13, the internal, private API of d3.map has changed
so that properties are not stored on the d3.map instance directly.  This
means that d3.values(d3.map({a: 1})) no longer returns [1].  Instead,
the public d3.map APIs should be used: d3.map({a: 1}).values().

I’ve also removed what looks like a redundant call to d3.values on an
array.