Distinguish empty page vs. no results returned
search_results now distinguishes betweeen when no query was entered (i.e., a
user going to the page without searching for anything) and when no results
were returned (e.g., a user searching for garbage). In addition, autocomplete
behavior was extended to the search bar on the 'results' page (which is on its
way to a proper landing page).