Refactor ajax for autosuggest to sequentially load data.
Previously, the location data ajax request and the full-text data ajax
request (for profiles, tables, and topics) were concurrently run within
a $.when(), then a $.then() was used to return the combined responses to
the autosuggest. Now a .then() chain is used to ensure that the location
data is loaded before the full-text data.