Style Explore search results to match Profile.
The Explore search results on the front page now have matching styles to
the Profile search results. The css was only accounting for having one
jQuery autocomplete box on the page (it looked for an id #ui-id-1; if
there are two, the second one will have an id of #ui-id-2).