Apply smaller autocomplete style only to header
Previously, the autocomplete results for the header search bar were
styled to have a smaller font size by adding a class to the autocomplete
li elements before as they were being rendered. However, the fact that
this code in full-text-results.js applied also to the search results
page search bar resulted in the full size search bar in the results page
having a smaller font size, which was not desired.

This issue was fixed by separating the autocomplete into two seaprate
jquery autocompletes, one being placed on the '#search-header' id and
the other being placed on the '#search' id (for the results page).